Refactor array classes to base module
Array-related classes, including `BaseArray` and `NamedArray`, were moved from `array.py` to `base.py` for better code organization. Corresponding import paths were also altered in `__init__.py` and `base.py`. This restructuring enhances the modularity and readability of the code.
