+++ src/mdio/schemas/__init__.py
@@ -1,7 +1,5 @@
 """MDIO schemas for different data types."""
 
-
-from mdio.schemas.array import NamedArray
 from mdio.schemas.compressors import ZFP
 from mdio.schemas.compressors import Blosc
 from mdio.schemas.dimension import NamedDimension
@@ -11,7 +9,6 @@ from mdio.schemas.dtype import StructuredType
 
 
 __all__ = [
-    "NamedArray",
     "Blosc",
     "ZFP",
     "NamedDimension",
