+++ src/mdio/schemas/base.py
@@ -1,10 +1,13 @@
 """Base models to subclass from."""
-
-
 from pydantic import BaseModel
 from pydantic import ConfigDict
+from pydantic import Field
 from pydantic.alias_generators import to_camel
 from pydantic.json_schema import GenerateJsonSchema
+
+from mdio.schemas import ZFP
+from mdio.schemas import Blosc
+from mdio.schemas import NamedDimension
 
 
 JSON_SCHEMA_DIALECT = GenerateJsonSchema.schema_dialect
@@ -25,3 +28,21 @@ class BaseDataset(StrictCamelBaseModel):
     """
 
     model_config = ConfigDict(json_schema_extra={"$schema": JSON_SCHEMA_DIALECT})
+
+
+class BaseArray(StrictCamelBaseModel):
+    """A base array schema."""
+
+    dimensions: list[NamedDimension] | list[str] = Field(
+        ..., description="List of Dimension collection or reference to dimension names."
+    )
+    compressor: Blosc | ZFP | None = Field(
+        default=None, description="Compression settings."
+    )
+
+
+class NamedArray(BaseArray):
+    """An array with a name."""
+
+    name: str = Field(..., description="Name of the array.")
+    long_name: str | None = Field(default=None, description="Fully descriptive name.")
